support experimental PR builds (#27)

+++ BuildWindowsRelease.sh
@@ -60,13 +60,18 @@ git pull
 
 # update OpenModelica
 cd /c/dev/${OM_ENCRYPT}OM${PLATFORM}
-git fetch && git fetch --tags
-
-# attempt to reset to origin
-git checkout "$OPENMODELICA_BRANCH" && git reset --hard "origin/$OPENMODELICA_BRANCH" || true
-git reset --hard "$OPENMODELICA_BRANCH" && git checkout "$OPENMODELICA_BRANCH" && git submodule update --force --init --recursive && git pull --recurse-submodules && git fetch && git fetch --tags || exit 1
-git checkout -f "$OPENMODELICA_BRANCH" || exit 1
-git reset --hard "$OPENMODELICA_BRANCH" || exit 1
+# if we build an experimental PR number skip this!
+if [[ "$1" =~ PR_.*_experimental ]]; then
+  echo "Building experimental PR build: $OPENMODELICA_BRANCH!"
+else
+  echo "Building $OPENMODELICA_BRANCH!"
+  git fetch && git fetch --tags
+  # attempt to reset to origin
+  git checkout "$OPENMODELICA_BRANCH" && git reset --hard "origin/$OPENMODELICA_BRANCH" || true
+  git reset --hard "$OPENMODELICA_BRANCH" && git checkout "$OPENMODELICA_BRANCH" && git submodule update --force --init --recursive && git pull --recurse-submodules && git fetch && git fetch --tags || exit 1
+  git checkout -f "$OPENMODELICA_BRANCH" || exit 1
+  git reset --hard "$OPENMODELICA_BRANCH" || exit 1
+fi
 git submodule update --force --init --recursive || exit 1
 
 # get the revision
